# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "השוואת נתוני תקופות עבודה - API",
    "version": "1.0.0",
    "description": "API סגור המשווה בין נתוני תקופות עבודה (משורות טבלת LD_CHINUCH_9050_TKUFOT_RETSIF) לבין דוח PDF \"סיכום נתוני פרישה\", עבור תעודת זהות אחת בכל קריאה (אחד-על-אחד): שורות ה-rows ומסמך ה-pdf חייבים להיות של אותה ת\"ז."
  },
  "servers": [{ "url": "/" }],
  "paths": {
    "/api/compare": {
      "post": {
        "summary": "השוואת שורות טבלה מול מסמך PDF (ת\"ז אחת)",
        "description": "אחד-על-אחד: rows חייב להכיל שורות של תעודת זהות אחת בלבד; בקשה עם יותר מת\"ז אחת נדחית עם 400. ניתן לשלוח בשתי צורות שקולות: JSON גולמי (pdf מקודד base64) או multipart/form-data (pdf כקובץ מצורף ממש, rows כשדה טקסט עם מערך JSON).",
        "parameters": [
          {
            "name": "full",
            "in": "query",
            "required": false,
            "description": "כשהערך '1', מוסיף לתשובה גם summary/warnings/results (הפירוט המלא שמסך התוצאות מציג). ברירת המחדל מחזירה תשובה רזה.",
            "schema": { "type": "string", "enum": ["1"] }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CompareRequestJson" },
              "example": {
                "rows": [
                  {
                    "MISPAR_TNUA": "9050",
                    "KOD_PEULA": "0",
                    "SEMEL_MISRAD": "0",
                    "MISPAR_ZEHUT": "012345678",
                    "ZIHUY_NOSAF": "0",
                    "SUG_TKUFA": "9999",
                    "TAARICH_ME": "01092010",
                    "TAARICH_AD": "31082015",
                    "ORECH_SHERUT": "06000",
                    "SUG_ZECHUYOT_LEGIMLA": "02",
                    "HEKEF_MISRA": "1000",
                    "SEQ": 1
                  }
                ],
                "pdf": {
                  "filename": "report_12345678.pdf",
                  "content": "<PDF בקידוד base64>"
                }
              }
            },
            "multipart/form-data": {
              "schema": { "$ref": "#/components/schemas/CompareRequestForm" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "ההשוואה בוצעה (גם כשהתוצאה אינה תקינה - valid=0 אינו שגיאת HTTP)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CompareResponse" }
              }
            }
          },
          "400": {
            "description": "בקשה לא תקינה: rows/pdf חסרים, יותר מת\"ז אחת ב-rows, או rows שאינו JSON תקין (במצב form-data)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/compare-summary": {
      "post": {
        "summary": "השוואת שורת \"סיכום תקופות עבודה - לפי אורך שירות\" (NETUNEY_TIK_MECHUSHAVIM) מול דוח PDF",
        "description": "משווה שורה אחת מהטבלה NETUNEY_TIK_MECHUSHAVIM (לפי אורך שירות בלבד, לא \"מחוץ לשירות\") מול טבלת \"סיכום תקופות עבודה\" שמודפסת באותו דוח \"סיכום נתוני פרישה\". בנוסף, בודק עצמאית שיחס השנים/חודשים בתוך הדוח עצמו תקין. ניתן לשלוח בשתי צורות שקולות: JSON גולמי (pdf מקודד base64) או multipart/form-data (pdf כקובץ מצורף ממש, row כשדה טקסט עם אובייקט JSON).",
        "parameters": [
          {
            "name": "full",
            "in": "query",
            "required": false,
            "description": "כשהערך '1', מוסיף לתשובה גם status/diffs/errors. ברירת המחדל מחזירה תשובה רזה (valid, text בלבד).",
            "schema": { "type": "string", "enum": ["1"] }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CompareSummaryRequestJson" },
              "example": {
                "row": {
                  "tkufa_mezaka_sherut": 132,
                  "chelkiyut_meshuklelet_sherut": 0.822,
                  "achuz_kizba_kafuf_chelkiyut": 22,
                  "achuz_kizb_achry_hagdl_chl_mla": 18.084
                },
                "pdf": {
                  "filename": "report_12345678.pdf",
                  "content": "<PDF בקידוד base64>"
                }
              }
            },
            "multipart/form-data": {
              "schema": { "$ref": "#/components/schemas/CompareSummaryRequestForm" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "ההשוואה בוצעה (גם כשהתוצאה אינה תקינה - valid=0 אינו שגיאת HTTP)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CompareSummaryResponse" }
              }
            }
          },
          "400": {
            "description": "בקשה לא תקינה: row/pdf חסרים, row עם שדה חסר/לא מספרי, או row שאינו JSON תקין (במצב form-data)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "TableRow": {
        "type": "object",
        "description": "שורת LD_CHINUCH_9050_TKUFOT_RETSIF; שמות העמודות כמפתחות, כפי שמוחזר מ-SELECT על הטבלה. עמודות עודפות (למשל LOAD_DATE) מותרות ומתעלמים מהן.",
        "additionalProperties": true,
        "properties": {
          "MISPAR_TNUA": { "type": "string", "example": "9050" },
          "MISPAR_ZEHUT": { "type": "string", "example": "012345678" },
          "SUG_TKUFA": { "type": "string", "example": "9999" },
          "TAARICH_ME": { "type": "string", "description": "DDMMYYYY", "example": "01092010" },
          "TAARICH_AD": { "type": "string", "description": "DDMMYYYY", "example": "31082015" },
          "ORECH_SHERUT": { "type": "string", "description": "חודשים כפול 100", "example": "06000" },
          "SUG_ZECHUYOT_LEGIMLA": { "type": "string", "example": "02" },
          "HEKEF_MISRA": { "type": "string", "description": "כפול 1000", "example": "1000" },
          "SEQ": { "type": "integer", "example": 1 }
        }
      },
      "PdfFile": {
        "type": "object",
        "required": ["filename", "content"],
        "properties": {
          "filename": { "type": "string", "example": "report_12345678.pdf" },
          "content": { "type": "string", "format": "byte", "description": "תוכן קובץ ה-PDF בקידוד base64" }
        }
      },
      "CompareRequestJson": {
        "type": "object",
        "required": ["rows", "pdf"],
        "properties": {
          "rows": { "type": "array", "items": { "$ref": "#/components/schemas/TableRow" } },
          "pdf": { "$ref": "#/components/schemas/PdfFile" }
        }
      },
      "CompareRequestForm": {
        "type": "object",
        "required": ["rows", "pdf"],
        "properties": {
          "rows": {
            "type": "string",
            "description": "מערך שורות הטבלה, מקודד כמחרוזת JSON",
            "example": "[{\"MISPAR_TNUA\":\"9050\",\"MISPAR_ZEHUT\":\"012345678\",\"SUG_TKUFA\":\"9999\",\"TAARICH_ME\":\"01092010\",\"TAARICH_AD\":\"31082015\",\"ORECH_SHERUT\":\"06000\",\"SUG_ZECHUYOT_LEGIMLA\":\"02\",\"HEKEF_MISRA\":\"1000\",\"SEQ\":1}]"
          },
          "pdf": { "type": "string", "format": "binary", "description": "קובץ ה-PDF, מצורף ממש" }
        }
      },
      "RowResult": {
        "type": "object",
        "description": "אותה שורה שנשלחה ב-rows, על כל שדותיה המקוריים, בתוספת:",
        "additionalProperties": true,
        "properties": {
          "valid": { "type": "integer", "enum": [0, 1], "description": "1 אם נמצאה תואמת במלואה במסמך, 0 אחרת" },
          "reason": { "type": "string", "description": "פירוט הפער כשאינה תקינה (מופיע רק כש-valid=0, או הערה כש-valid=1 למשל שורת עזיבה)" }
        }
      },
      "CompareResponse": {
        "type": "object",
        "properties": {
          "valid": { "type": "integer", "enum": [0, 1], "description": "1 רק כשכל התקופות תואמות במלואן" },
          "idNumber": { "type": ["string", "null"], "example": "12345678" },
          "rows": { "type": "array", "items": { "$ref": "#/components/schemas/RowResult" } },
          "text": { "type": "string", "description": "טקסט מאוחד המרכז את כל התוצאות והפערים" },
          "summary": {
            "type": "object",
            "description": "מוחזר רק כש-?full=1",
            "properties": {
              "total": { "type": "integer" },
              "match": { "type": "integer" },
              "mismatch": { "type": "integer" },
              "missing": { "type": "integer" },
              "error": { "type": "integer" }
            }
          },
          "warnings": { "type": "array", "items": { "type": "string" }, "description": "מוחזר רק כש-?full=1" },
          "results": { "type": "array", "items": { "type": "object" }, "description": "מוחזר רק כש-?full=1 - הפירוט המלא (מה שמסך התוצאות מציג)" }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        }
      },
      "DbWorkSummaryRow": {
        "type": "object",
        "description": "שורת NETUNEY_TIK_MECHUSHAVIM (לפי אורך שירות בלבד - אין עמודת \"שנים\", היא נגזרת מהחודשים).",
        "required": ["tkufa_mezaka_sherut", "chelkiyut_meshuklelet_sherut", "achuz_kizba_kafuf_chelkiyut", "achuz_kizb_achry_hagdl_chl_mla"],
        "properties": {
          "tkufa_mezaka_sherut": { "type": "number", "description": "סך חודשים", "example": 132 },
          "chelkiyut_meshuklelet_sherut": { "type": "number", "description": "חלקיות משוקללת", "example": 0.822 },
          "achuz_kizba_kafuf_chelkiyut": { "type": "number", "description": "אחוז קצבה כפוף לחלקיות", "example": 22 },
          "achuz_kizb_achry_hagdl_chl_mla": { "type": "number", "description": "אחוז לפי חלקיות מלאה", "example": 18.084 }
        }
      },
      "CompareSummaryRequestJson": {
        "type": "object",
        "required": ["row", "pdf"],
        "properties": {
          "row": { "$ref": "#/components/schemas/DbWorkSummaryRow" },
          "pdf": { "$ref": "#/components/schemas/PdfFile" }
        }
      },
      "CompareSummaryRequestForm": {
        "type": "object",
        "required": ["row", "pdf"],
        "properties": {
          "row": {
            "type": "string",
            "description": "שורת ה-DB, מקודדת כמחרוזת JSON",
            "example": "{\"tkufa_mezaka_sherut\":132,\"chelkiyut_meshuklelet_sherut\":0.822,\"achuz_kizba_kafuf_chelkiyut\":22,\"achuz_kizb_achry_hagdl_chl_mla\":18.084}"
          },
          "pdf": { "type": "string", "format": "binary", "description": "קובץ ה-PDF, מצורף ממש" }
        }
      },
      "CompareSummaryResponse": {
        "type": "object",
        "properties": {
          "valid": { "type": "integer", "enum": [0, 1], "description": "1 רק כשהשורה תואמת במלואה (status=match)" },
          "text": { "type": "string", "description": "טקסט מאוחד המרכז את התוצאה והפערים" },
          "status": {
            "type": "string",
            "enum": ["match", "mismatch", "missing_pdf", "error"],
            "description": "מוחזר רק כש-?full=1"
          },
          "diffs": {
            "type": "array",
            "items": { "type": "object" },
            "description": "מוחזר רק כש-?full=1 - פירוט השדות שאינם תואמים"
          },
          "errors": {
            "type": "array",
            "items": { "type": "string" },
            "description": "מוחזר רק כש-?full=1"
          }
        }
      }
    }
  }
}